Changed some 'int's to 'size_t's, to improve 64-bit compatibility. Also changed jmin and jmax to use templates, so they can take any type. These changes might mean that you'll need to add some more explicit casts to get your existing code to compile, but this is actually a good thing - it brought to light a few dodgy implicit casts in my code, and may do the same in yours. Also added a function roundToInt(), which replaces roundDoubleToInt() and roundFloatToInt(), but takes any size of floating point number (I've left the old roundDoubleToInt() and roundFloatToInt() functions there for convenience, but will probably remove them in the future).

+++ src/audio/audio_file_formats/juce_FlacAudioFormat.cpp
@@ -328,7 +328,7 @@ public:
         FLAC__stream_encoder_set_do_mid_side_stereo (encoder, numChannels == 2);
         FLAC__stream_encoder_set_loose_mid_side_stereo (encoder, numChannels == 2);
         FLAC__stream_encoder_set_channels (encoder, numChannels);
-        FLAC__stream_encoder_set_bits_per_sample (encoder, jmin (24, bitsPerSample));
+        FLAC__stream_encoder_set_bits_per_sample (encoder, jmin ((unsigned int) 24, bitsPerSample));
         FLAC__stream_encoder_set_sample_rate (encoder, (unsigned int) sampleRate);
         FLAC__stream_encoder_set_blocksize (encoder, 2048);
         FLAC__stream_encoder_set_do_escape_coding (encoder, true);
